refactor snapshot logic into shared module
