+++ js/competitions.js
@@ -1,3 +1,5 @@
+import { captureSnapshots } from './snapshot.js';
+
 const API_BASE = (window.API_ORIGIN || '') + '/api';
 
 function like(id) {
@@ -17,33 +19,6 @@ function like(id) {
     });
 }
 
-async function captureSnapshots(container) {
-  const cards = container.querySelectorAll('.entry-card');
-  for (const card of cards) {
-    const img = card.querySelector('img');
-    if (img && img.src) continue;
-    const glbUrl = card.dataset.model;
-    const viewer = document.createElement('model-viewer');
-    viewer.src = glbUrl;
-    viewer.setAttribute(
-      'environment-image',
-      'https://modelviewer.dev/shared-assets/environments/neutral.hdr'
-    );
-    viewer.style.position = 'fixed';
-    viewer.style.left = '-10000px';
-    viewer.style.width = '300px';
-    viewer.style.height = '300px';
-    document.body.appendChild(viewer);
-    try {
-      await viewer.updateComplete;
-      img.src = await viewer.toDataURL('image/png');
-    } catch (err) {
-      console.error('Failed to capture snapshot', err);
-    } finally {
-      viewer.remove();
-    }
-  }
-}
 
 async function load() {
   const res = await fetch(`${API_BASE}/competitions/active`);
